# bicep/ag-apim-aks-simplified/src/order/openapi.yaml
---
openapi: 3.0.2
info:
  title: Order Api
  version: 0.1.0
paths:
  "/swagger":
    get:
      summary: Get Swagger
      operationId: getSwagger
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Orders
                type: object
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  "/orders":
    get:
      summary: Get Orders
      operationId: getOrders
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Orders
                type: array
                items:
                  "$ref": "#/components/schemas/Order"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: New Order
      operationId: newOrder
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/NewOrder"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Order"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  "/orders/{id}":
    get:
      summary: Get Order
      operationId: getOrder
      parameters:
        - required: true
          schema:
            title: id
            type: integer
          name: id
          in: path
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Order"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update Order
      operationId: updateOrder
      parameters:
        - required: true
          schema:
            title: id
            type: integer
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/NewOrder"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Order"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete Order
      operationId: deleteOrder
      parameters:
        - required: true
          schema:
            title: id
            type: integer
          name: id
          in: path
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            "$ref": "#/components/schemas/ValidationError"
    Order:
      title: Order
      required:
        - Name
        - ID
      type: object
      properties:
        Name:
          title: Name
          type: string
        Description:
          title: Description
          type: string
        ID:
          title: ID
          type: integer
    NewOrder:
      title: NewOrder
      required:
        - Name
      type: object
      properties:
        Name:
          title: Name
          type: string
        Description:
          title: Description
          type: string
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
    Error:
      title: Error
      required:
        - code
        - message
      type: object
      properties:
        code:
          type: integer
          description: Error code
          format: int32
        message:
          type: string
          description: Error message